refactor: created usecase authentication domain
