+++ app/application/usecase/authentication/login_user.go
@@ -1,4 +1,4 @@
-package account
+package authentication
 
 import (
 	"context"
