+++ app/application/usecase/authentication/login_user_test.go
@@ -1,4 +1,4 @@
-package account
+package authentication
 
 import (
 	"context"
